# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exam generator API",
        "description": "API for exam generator project at UNSA",
        "contact": {
            "email": "rhualla@unsa.edu.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "/var/www/html/exam_generator/backend/app"
        }
    ],
    "paths": {
...
        "/processes": {
            "get": {
                "tags": [
                    "Procesos"
                ],
                "summary": "Obtener lista de procesos",
                "description": "Retorna una lista de procesos",
                "operationId": "014b44c37d00ed39f0b2b8009eb0084c",
                "responses": {
                    "200": {
                        "description": "Lista de procesos"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Procesos"
                ],
                "summary": "Crear un nuevo proceso",
                "description": "Crea un nuevo proceso y lo devuelve",
                "operationId": "9ec02bb1d645b177be4ca7a7f5ad6ac9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Proceso de admisión"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proceso creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Proceso de admisión"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-27T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-27T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            }
        },
        "/processes/{id}": {
            "get": {
                "tags": [
                    "Procesos"
                ],
                "summary": "Obtiene informacion de un proceso",
                "description": "Retorna datos de un proceso",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del proceso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Process data"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Procesos"
                ],
                "summary": "Eliminar un proceso",
                "description": "Elimina un proceso por su ID",
                "operationId": "4f3ed55cbed755f03e664bf4f5f34736",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proceso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Proceso eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proceso no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Procesos"
                ],
                "summary": "Actualizar un proceso",
                "description": "Actualiza los datos de un proceso existente",
                "operationId": "d970ffcc85e9a41c034285d78f3f220c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proceso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Proceso actualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proceso actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Proceso actualizado"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name must not be greater than 255 characters."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proceso no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum_token": []
                    }
                ]
            }
        },
...
    },
...
}